requirements.txt
matplotlib==3.6.3
jupyterlab==3.5.3
scipy==1.10.0
scikit-learn==1.2.1
scikit-image==0.19.3
pandas==1.5.3
music21==8.1.0
black[jupyter]==23.1.0
click==8.0.4
flake8==5.0.4
kaggle==1.5.12
pydot==1.4.2
ipywidgets==8.0.4
tensorflow-macos==2.10.0
tensorboard==2.10.1
tensorflow_probability==0.18.0
flake8-nb==0.5.2